# lib/api-zod/src/generated/types/adminInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Moats App Governance API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminInput {
  projectId: number;
  walletAddress: string;
}