# lib/api-zod/src/generated/types/proposalInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Moats App Governance API
 * OpenAPI spec version: 0.1.0
 */
import type { ProposalInputQuorumType } from './proposalInputQuorumType';

export interface ProposalInput {
  title: string;
  description?: string;
  projectId: number;
  quorumType: ProposalInputQuorumType;
  quorumThreshold: number;
  startDate: string;
  endDate: string;
  createdBy: string;
}