# lib/api-zod/src/generated/types/proposalSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Moats App Governance API
 * OpenAPI spec version: 0.1.0
 */

export interface ProposalSummary {
  total: number;
  active: number;
  passed: number;
  failed: number;
  pending: number;
  cancelled: number;
}