# lib/api-zod/src/generated/types/proposalUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Moats App Governance API
 * OpenAPI spec version: 0.1.0
 */
import type { ProposalUpdateStatus } from './proposalUpdateStatus';

export interface ProposalUpdate {
  status?: ProposalUpdateStatus;
  title?: string;
  description?: string;
  endDate?: string;
}